renamed test file to metric
